Support fractional scaling.

+++ hw/xwayland/xwayland-screen.h
@@ -86,7 +86,7 @@ struct xwl_screen {
     struct xorg_list damage_window_list;
     struct xorg_list window_list;
 
-    int32_t global_output_scale;
+    double global_output_scale;
 
     int wayland_fd;
     struct wl_display *display;
@@ -166,7 +166,7 @@ void xwl_surface_damage(struct xwl_screen *xwl_screen,
                         int32_t x, int32_t y, int32_t width, int32_t height);
 int xwl_screen_get_next_output_serial(struct xwl_screen * xwl_screen);
 
-int xwl_scale_to(struct xwl_screen *xwl_screen, int value);
-void xwl_screen_set_global_scale(struct xwl_screen *xwl_screen, int32_t scale);
+double xwl_scale_to(struct xwl_screen *xwl_screen, double value);
+void xwl_screen_set_global_scale(struct xwl_screen *xwl_screen, double scale);
 
 #endif /* XWAYLAND_SCREEN_H */
